requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
sqlalchemy
chromadb
tiktoken
cryptography
httpx
apscheduler
google-generativeai
google-auth
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai
pypdf
psycopg2-binary
email-validator
requests
python-dotenv
